Modify main class and add clean_up module
Move the create cache directory and connect to solr from the initialize
method in the main class to their own method -> set_up. Call set_up in
the right place. This allows the cean_up methods to not have to create a
cache dir and connect to solr when run.

This also adds a task for cleaning up the cache directory (basically
deleting it) and sets the default task to run the clone task.